explicitly calling dplyr::select
changes to payroll gathering:

in salary DF, changed 'value' to a more descriptive 'salary'

not returning empty or FA years as they're no longer needed. doing this also ensures that 1 row = 1 year of control in all_teams_payroll.

in the payroll DF returned, taking only up to the next 5 years for each player.