got pretty far into a tidy data conversion.
but I think my salaries are too high. guys like bryant are showing up with $80m of salaries for their arb years. this may be OK because he's making $10m this year.

but guys like wade leblanc are crazy high too.

and all teams come out with negative surplus value. this means my projections are too low or the salaries are too high.