i realized with a shock that a simple linear model works really, really well if the player has already been on the ballot once.
so i added code that does that.

so i think all I need ML to do is to try and predict the first-year players.

to do that: need to split into 3 types (position, starter, reliever) and go from there.

but a simple combination of BJ HOF score + narrative score gets us a decent way there for a lot of players.